app/requirements.txt
SQLAlchemy==1.3.11
sqlalchemy-migrate==0.13.0
Telethon==1.32.0
mysql-connector-python==8.0.18
gspread==3.1.0
oauth2client==4.1.3
python-dotenv==0.19.0
elasticsearch==7.13.1
uritools==4.0.1
urlextract==1.8.0
pyyaml==6.0.1
langdetect==1.0.9
geotext==0.4.0
jionlp==1.5.2
